Rename package bitbucket to bitbucketcloud
